Moved the user modal, plus data sanity
