--- src/components/InfoAwards.vue
@@ -1,10 +1,0 @@
-<script setup>
-
-import AwardView from './award/AwardView.vue';
-
-</script>
-
-
-<template>
-<AwardView></AwardView>
-</template>
