--- src/components/InfoEducation.vue
@@ -1,10 +1,0 @@
-<script setup>
-
-import EducationView from './education/EducationView.vue';
-
-</script>
-
-
-<template>
-<EducationView></EducationView>
-</template>
