--- src/components/InfoExperience.vue
@@ -1,10 +1,0 @@
-<script setup>
-
-import ExperienceView from './experience/ExperienceView.vue';
-
-</script>
-
-
-<template>
-<ExperienceView></ExperienceView>
-</template>
