--- src/components/InfoLinks.vue
@@ -1,13 +1,0 @@
-<script setup>
-
-import LinksView from './links/LinkView.vue';
-
-
-</script>
-
-
-<template>
-
-<LinksView></LinksView>
-
-</template>
