--- src/components/InfoProjects.vue
@@ -1,10 +1,0 @@
-<script setup>
-
-import ProjectView from './project/ProjectView.vue';
-
-</script>
-
-
-<template>
-<ProjectView></ProjectView>
-</template>
