--- src/components/InfoSkills.vue
@@ -1,10 +1,0 @@
-<script setup>
-
-import SkillView from './skill/SkillView.vue';
-
-</script>
-
-
-<template>
-<SkillView></SkillView>
-</template>
