--- src/components/InfoUsers.vue
@@ -1,15 +1,0 @@
-<script setup>
-
-import UserView from './user/UserView.vue';
-
-</script>
-
-
-<template>
-
-<UserView></UserView>
-
-</template>
-
-<style scoped>
-</style>
